+++ CHANGELOG.md
@@ -1,3 +1,24 @@
+# [1.391.0](https://github.com/mkayander/dStruct/compare/v1.390.0...v1.391.0) (2026-09-03)
+
+
+### Features
+
+* **app:** migrate API routes to App Router (P7–P8) ([#182](https://github.com/mkayander/dStruct/issues/182)) ([f25c7cd](https://github.com/mkayander/dStruct/commit/f25c7cd1edae0a7ff6b320d3f42979b02f5c7daf))
+
+# [1.390.0](https://github.com/mkayander/dStruct/compare/v1.389.0...v1.390.0) (2026-09-03)
+
+
+### Features
+
+* **app:** migrate sitemap to App Router (P6) ([#181](https://github.com/mkayander/dStruct/issues/181)) ([f442d27](https://github.com/mkayander/dStruct/commit/f442d27b4daaa08df2ca0debd18078d0b1b10c4b))
+
+# [1.389.0](https://github.com/mkayander/dStruct/compare/v1.388.0...v1.389.0) (2026-09-03)
+
+
+### Features
+
+* **app:** L5 Suspense device hints + playground instant adoption ([#179](https://github.com/mkayander/dStruct/issues/179)) ([cc40278](https://github.com/mkayander/dStruct/commit/cc40278b50c6c36e05c2567c1d0bd44043be5646))
+
 # [1.388.0](https://github.com/mkayander/dStruct/compare/v1.387.0...v1.388.0) (2026-08-29)